Refactor navigation files
