+++ src/navigation/StackNavigation.js
@@ -1,21 +1,21 @@
 import { createStackNavigator } from "@react-navigation/stack";
-import { THEME } from "../../constants/Theme";
-import BacktestingScreen from "../main_screens/BacktestingScreen";
-import ProfileScreen from "../main_screens/ProfileScreen";
-import HomeScreen from "../main_screens/home/HomeScreen";
-import CreateInvestorStep1Screen from "../main_screens/home/create_investor/CreateInvestorStep1Screen";
-import CreateInvestorAIStep2Screen from "../main_screens/home/create_investor/ai/CreateInvestorAIStep2Screen";
-import CreateInvestorAIStep3Screen from "../main_screens/home/create_investor/ai/CreateInvestorAIStep3Screen";
-import CreateInvestorAlgorithmicStep2Screen from "../main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep2Screen";
-import CreateInvestorAlgorithmicStep3Screen from "../main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep3Screen";
-import CreateInvestorAlgorithmicStep4Screen from "../main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep4Screen";
-import InvestorScreen from "../main_screens/home/invest/InvestorScreen";
-import JobScreen from "../main_screens/home/invest/JobScreen";
-import SearchScreen from "../main_screens/search/SearchScreen";
-import StockInfoScreen from "../main_screens/search/results/StockInfoScreen";
-import BacktestResultsScreen from "../nested_screens/backtesting/BacktestResultsScreen";
-import CreateBacktestScreen from "../nested_screens/backtesting/CreateBacktestScreen";
-import TradeHistoryScreen from "../nested_screens/profile/TradeHistoryScreen";
+import BacktestingScreen from "../components/main_screens/BacktestingScreen";
+import ProfileScreen from "../components/main_screens/ProfileScreen";
+import HomeScreen from "../components/main_screens/home/HomeScreen";
+import CreateInvestorStep1Screen from "../components/main_screens/home/create_investor/CreateInvestorStep1Screen";
+import CreateInvestorAIStep2Screen from "../components/main_screens/home/create_investor/ai/CreateInvestorAIStep2Screen";
+import CreateInvestorAIStep3Screen from "../components/main_screens/home/create_investor/ai/CreateInvestorAIStep3Screen";
+import CreateInvestorAlgorithmicStep2Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep2Screen";
+import CreateInvestorAlgorithmicStep3Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep3Screen";
+import CreateInvestorAlgorithmicStep4Screen from "../components/main_screens/home/create_investor/algorithmic/CreateInvestorAlgorithmicStep4Screen";
+import InvestorScreen from "../components/main_screens/home/invest/InvestorScreen";
+import JobScreen from "../components/main_screens/home/invest/JobScreen";
+import SearchScreen from "../components/main_screens/search/SearchScreen";
+import StockInfoScreen from "../components/main_screens/search/results/StockInfoScreen";
+import BacktestResultsScreen from "../components/nested_screens/backtesting/BacktestResultsScreen";
+import CreateBacktestScreen from "../components/nested_screens/backtesting/CreateBacktestScreen";
+import TradeHistoryScreen from "../components/nested_screens/profile/TradeHistoryScreen";
+import { THEME } from "../constants/Theme";
 import { nestedScreenOptions } from "./NavigationHelpers";
 
 const Stack = createStackNavigator();
